Break out snapshots from FSMClient with FSMClient.NewSnapshotter bridge
